pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel", "Cython>3.1", "numpy"]
build-backend = "setuptools.build_meta"

[project]
name = "parallel-sparse-tools"
version = "0.2.5"
authors = [
    {name = "Phillip Weinberg", email = "weinbe58@gmail.com"},
]
description = "My package description"
readme = "README.md"
requires-python = ">=3.11,<3.15"
license = {text = "BSD-3-Clause"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "numpy>=2.0.0",
    "scipy>=1.13.0",
]


[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.3.1",
]
slepc = [
    "petsc4py",
    "slepc4py"
]

[project.optional-dependencies]
slepc = [
    "mpi4py",
    "petsc4py",
    "slepc4py"
]
 
[tool.pdm.scripts]
_.env_file = "./.env"
upload = "python -m twine upload -r bloqade dist/*"
build = {cmd = "pdm build"}
publish = {composite = ["build", "upload"]}
test = {shell = "pytest -x"}
install-slepc = {call = "slepc_install.install_slepc:main", env = {PETSC_CONFIGURE_OPTIONS = "--download-f2cblaslapack --with-debugging=0"}}
[tool.cibuildwheel]
skip = ""

build-frontend = "default"
config-settings = {}
dependency-versions = "pinned"
#environment = {}
environment-pass = []
build-verbosity = 1


before-build = ""
repair-wheel-command = ""

before-test = ""
test-command = "pytest {project}/test"
test-requires = ["pytest"]
test-extras = ["slepc"]


container-engine = "docker"

manylinux-x86_64-image = "manylinux_2_28"
manylinux-i686-image = "manylinux_2_28"
manylinux-aarch64-image = "manylinux_2_28"
manylinux-ppc64le-image = "manylinux_2_28"
manylinux-s390x-image = "manylinux_2_28"
manylinux-pypy_x86_64-image = "manylinux_2_28"
manylinux-pypy_i686-image = "manylinux_2_28"
manylinux-pypy_aarch64-image = "manylinux_2_28"

musllinux-x86_64-image = "musllinux_1_2"
musllinux-i686-image = "musllinux_1_2"
musllinux-aarch64-image = "musllinux_1_2"
musllinux-ppc64le-image = "musllinux_1_2"
musllinux-s390x-image = "musllinux_1_2"


[tool.cibuildwheel.linux]
before-all = "PETSC_CONFIGURE_OPTIONS='--with-mpi=0 --download-f2cblaslapack' pip install petsc petsc4py slepc slepc4py"
archs = ["native"]
repair-wheel-command = "auditwheel repair -w {dest_dir} {wheel}"

[tool.cibuildwheel.macos]
before-all = "PETSC_CONFIGURE_OPTIONS='--with-mpi=0 --download-f2cblaslapack' pip install petsc petsc4py slepc slepc4py"
archs = ["native"]
repair-wheel-command = "delocate-wheel --exclude libomp --require-archs {delocate_archs} -w {dest_dir} -v {wheel}"


[tool.cibuildwheel.windows]
archs = ["native"]
